fix: 统一数据通路 + review 问题修复
- P0: AppInfoProvider 改读 Room 内部 fakegps.db，解决双库分裂导致 hook 读不到新位置
- P1: nci 类型从 Int 改为 Long (36-bit NR Cell Identity)
- P1: MyLocation 按钮接通 permLauncher 权限请求链
- P2: 移除 0.0 坐标守卫，赤道/零子午线位置不再被吞
- P2: VerifyViewModel cursor 泄漏修复 (cursor.use)
- 新增: 验证页"数据通道"自检 (Room vs ContentProvider 一致性对比)

+++ app/src/main/java/name/caiyao/fakegps/data/db/ProfileEntity.kt
@@ -43,7 +43,7 @@ data class ProfileEntity(
     @ColumnInfo(name = "lte_bandwidth") val lteBandwidth: Int? = null,
 
     // Cell Identity — NR/5G
-    val nci: Int? = null,
+    val nci: Long? = null,
     val nrarfcn: Int? = null,
     @ColumnInfo(name = "nr_pci") val nrPci: Int? = null,
     @ColumnInfo(name = "nr_tac") val nrTac: Int? = null,
